+++ packages/ui/Slider/AbstractSliderChild.ts
@@ -1,7 +1,7 @@
 import { Base } from '@studiometa/js-toolkit';
-import type { BaseProps, BaseConfig, BaseInterface } from '@studiometa/js-toolkit';
+import type { BaseProps, BaseConfig } from '@studiometa/js-toolkit';
 import { nextFrame, domScheduler, isFunction } from '@studiometa/js-toolkit/utils';
-import { Slider } from './Slider.js';
+import type { Slider } from './Slider.js';
 
 export interface AbstractSliderChildProps extends BaseProps {
   $parent: Slider;
@@ -10,10 +10,9 @@ export interface AbstractSliderChildProps extends BaseProps {
 /**
  * AbstractSliderChild class.
  */
-export class AbstractSliderChild<T extends BaseProps = BaseProps>
-  extends Base<T & AbstractSliderChildProps>
-  implements BaseInterface
-{
+export class AbstractSliderChild<T extends BaseProps = BaseProps> extends Base<
+  T & AbstractSliderChildProps
+> {
   /**
    * Config.
    */
@@ -25,12 +24,6 @@ export class AbstractSliderChild<T extends BaseProps = BaseProps>
    * Listen to the `goto` event of the parent on mount.
    */
   mounted() {
-    if (!(this.$parent instanceof Slider)) {
-      throw new Error(
-        `The \`${this.$options.name}\` component must be a direct child of a \`Slider\` component.`,
-      );
-    }
-
     this.$parent.$on('index', this);
   }
 
@@ -52,10 +45,8 @@ export class AbstractSliderChild<T extends BaseProps = BaseProps>
 
   /**
    * Dispatch event.
-   * @param   {CustomEvent} event
-   * @returns {void}
    */
-  handleEvent(event) {
+  handleEvent(event: CustomEvent) {
     if (event.type === 'index') {
       domScheduler.read(() => {
         const callback = this.update(event.detail[0]);
@@ -69,13 +60,10 @@ export class AbstractSliderChild<T extends BaseProps = BaseProps>
     }
   }
 
-  // eslint-disable-next-line jsdoc/require-returns-check
   /**
    * Update the child component with the given index.
-   * @param   {number} index The new active index.
-   * @returns {void|(()=>void)}
    */
-  update(index): void | (() => void) {
+  update(index: number): void | (() => void) {
     throw new Error(`The \`AbstractSliderChild.update(${index})\` method must be implemented.`);
   }
 }